make query digesting standard pipelines work with pipeline.eval()
