+++ haystack/pipelines/standard_pipelines.py
@@ -70,6 +70,27 @@ class BaseStandardPipeline(ABC):
         """
         self.pipeline.draw(path)
 
+    def eval(self,
+            queries: List[str],
+            labels: List[MultiLabel],
+            params: Optional[dict],
+            sas_model_name_or_path: str = None) -> EvaluationResult:
+            
+        """
+        Evaluates the pipeline by running the pipeline once per query in debug mode 
+        and putting together all data that is needed for evaluation, e.g. calculating metrics.
+
+        :param queries: The queries to evaluate
+        :param labels: The labels to evaluate on
+        :param params: Params for the `retriever` and `reader`. For instance,
+                       params={"Retriever": {"top_k": 10}, "Reader": {"top_k": 5}}
+        :param sas_model_name_or_path: SentenceTransformers semantic textual similarity model to be used for sas value calculation, 
+                                    should be path or string pointing to downloadable models.
+        """
+        output = self.pipeline.eval(queries=queries, labels=labels, params=params, 
+            sas_model_name_or_path=sas_model_name_or_path)
+        return output
+
 
 class ExtractiveQAPipeline(BaseStandardPipeline):
     """
@@ -99,27 +120,6 @@ class ExtractiveQAPipeline(BaseStandardPipeline):
                       by this method under the key "_debug"
         """
         output = self.pipeline.run(query=query, params=params, debug=debug)
-        return output
-
-    def eval(self,
-            queries: List[str],
-            labels: List[MultiLabel],
-            params: Optional[dict],
-            sas_model_name_or_path: str = None) -> EvaluationResult:
-            
-        """
-        Evaluates the pipeline by running the pipeline once per query in debug mode 
-        and putting together all data that is needed for evaluation, e.g. calculating metrics.
-
-        :param queries: The queries to evaluate
-        :param labels: The labels to evaluate on
-        :param params: Params for the `retriever` and `reader`. For instance,
-                       params={"Retriever": {"top_k": 10}, "Reader": {"top_k": 5}}
-        :param sas_model_name_or_path: SentenceTransformers semantic textual similarity model to be used for sas value calculation, 
-                                    should be path or string pointing to downloadable models.
-        """
-        output = self.pipeline.eval(queries=queries, labels=labels, params=params, 
-            sas_model_name_or_path=sas_model_name_or_path)
         return output
 
     def _format_document_answer(self, document_or_answer: dict):
